SMOKEZIP: add Author, Revision and Date keywords to smokezip source
Former-commit-id: 78c4602635de2dc12bc5513fbbc551075c77fc6c [formerly c3f20ffb78f1bd89b8f02d440102a70979bc7836]
Former-commit-id: 56fc275b839aacb5add21fab0b2cb218687cc321
Former-commit-id: 70d14c7c7d07bb3528b84428a57a7b838b31cb96

+++ SMV_5/source/smokezip/CNVboundary.c
@@ -1,3 +1,7 @@
+// $Date$ 
+// $Revision$
+// $Author$
+
 #include "options.h"
 #include "zlib.h"
 #include <stdio.h>
@@ -313,7 +317,7 @@ int convert_boundary(patch *patchi,int pass){
       }
 //      printf("time=%f before %i after=%i\n",time,npatchfull,ncompressed_zlib);
 
-      fprintf(boundarysizestream,"%f %i %i\n",time,npatchfull,ncompressed_zlib);
+      fprintf(boundarysizestream,"%f %i %i\n",time,(int)npatchfull,(int)ncompressed_zlib);
       fwrite(&time,4,1,boundarystream);                                       // write out time
       fwrite(&ncompressed_zlib,4,1,boundarystream);                           // write out compressed size of frame
       fwrite(compressed_boundarybuffer,1,ncompressed_zlib,boundarystream);    // write out compressed buffer
